+++ changelog.md
@@ -1,2 +1,1 @@
-- fixed biome modifiers ambient and music sounds not decoding properly
-- fixed color diffuse color having alpha set to 0
+- fixed an issue with custom creative tabs